A filter-based implementation for <5.12.
I wanted to provide a proper rewrite of lex_stuff, but it doesn't seem
to be possible. rafl's B::Hooks::Parser does the best that can be done,
which means using a source filter to preallocate space in PL_linestr.
This is necessary since before lex_stuff_sv, bits of toke.c used to hold
on to pointers into PL_linestr, so it can't be reallocated.

Unfortunately, this means BHP needs you to call BHP::setup at least one
source line before the point where the text is to be inserted, to give
the filter a chance to run. This is exactly what I cannot promise to do,
in this case, so it's not usable.

This implementation just uses a filter directly, which has two caveats:
first, the insertion happens at the start of the next line rather than
immediately; and second, text cannot be inserted into string evals.
String evals are compiled under the assumption that the string has
already been filtered (when its surrounding file was read), so filters
are not applied.

It's important to catch attempts to insert into an eval, rather than
letting them pass, as otherwise the text gets inserted into some random
other file. The context immediately outside the innermost BEGIN is an
eval {}; the next context is the one currently being compiled.